+++ tests/cli/env/test_prune.py
@@ -1,28 +1,26 @@
-from hatch.config.constants import AppEnvVars, ConfigEnvVars
+from hatch.config.constants import AppEnvVars
 from hatch.project.core import Project
 
 
-def test_unknown_type(hatch, helpers, temp_dir, config_file):
+def test_unknown_type(hatch, helpers, temp_dir_data, config_file):
     config_file.model.template.plugins['default']['tests'] = False
     config_file.save()
 
     project_name = 'My App'
 
-    with temp_dir.as_cwd():
+    with temp_dir_data.as_cwd():
         result = hatch('new', project_name)
 
     assert result.exit_code == 0, result.output
 
-    project_path = temp_dir / 'my-app'
-    cache_path = temp_dir / 'cache'
-    cache_path.mkdir()
+    project_path = temp_dir_data / 'my-app'
 
     project = Project(project_path)
     config = dict(project.config.envs['default'])
     config['type'] = 'foo'
     helpers.update_project_environment(project, 'default', config)
 
-    with project_path.as_cwd(env_vars={ConfigEnvVars.CACHE: str(cache_path)}):
+    with project_path.as_cwd():
         result = hatch('env', 'prune')
 
     assert result.exit_code == 1
@@ -33,19 +31,15 @@ def test_unknown_type(hatch, helpers, temp_dir, config_file):
     )
 
 
-def test_all(hatch, helpers, temp_dir, config_file):
+def test_all(hatch, helpers, temp_dir_data, config_file):
     project_name = 'My App'
 
-    cache_path = temp_dir / 'cache'
-    config_file.model.dirs.env = str(cache_path)
-    config_file.save()
-
-    with temp_dir.as_cwd():
+    with temp_dir_data.as_cwd():
         result = hatch('new', project_name)
 
     assert result.exit_code == 0, result.output
 
-    project_path = temp_dir / 'my-app'
+    project_path = temp_dir_data / 'my-app'
 
     project = Project(project_path)
     helpers.update_project_environment(project, 'default', {'skip-install': True, **project.config.envs['default']})
@@ -62,7 +56,7 @@ def test_all(hatch, helpers, temp_dir, config_file):
 
     assert result.exit_code == 0, result.output
 
-    env_cache_path = cache_path / 'env' / 'virtual'
+    env_cache_path = temp_dir_data / 'data' / 'env' / 'virtual'
     assert env_cache_path.is_dir()
 
     storage_dirs = list(env_cache_path.iterdir())
@@ -88,19 +82,15 @@ def test_all(hatch, helpers, temp_dir, config_file):
     assert not storage_path.is_dir()
 
 
-def test_incompatible_ok(hatch, helpers, temp_dir, config_file):
+def test_incompatible_ok(hatch, helpers, temp_dir_data, config_file):
     project_name = 'My App'
 
-    cache_path = temp_dir / 'cache'
-    config_file.model.dirs.env = str(cache_path)
-    config_file.save()
-
-    with temp_dir.as_cwd():
+    with temp_dir_data.as_cwd():
         result = hatch('new', project_name)
 
     assert result.exit_code == 0, result.output
 
-    project_path = temp_dir / 'my-app'
+    project_path = temp_dir_data / 'my-app'
 
     project = Project(project_path)
     helpers.update_project_environment(
@@ -114,19 +104,15 @@ def test_incompatible_ok(hatch, helpers, temp_dir, config_file):
     assert not result.output
 
 
-def test_active(hatch, temp_dir, helpers, config_file):
+def test_active(hatch, temp_dir_data, helpers, config_file):
     project_name = 'My App'
 
-    cache_path = temp_dir / 'cache'
-    config_file.model.dirs.env = str(cache_path)
-    config_file.save()
-
-    with temp_dir.as_cwd():
+    with temp_dir_data.as_cwd():
         result = hatch('new', project_name)
 
     assert result.exit_code == 0, result.output
 
-    project_path = temp_dir / 'my-app'
+    project_path = temp_dir_data / 'my-app'
 
     with project_path.as_cwd(env_vars={AppEnvVars.ENV_ACTIVE: 'default'}):
         result = hatch('env', 'prune')