+++ docs/meta/history.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Change environment storage configuration to allow for exact paths per type of environment
 
+***Added:***
+
+- Add more informative help text for the arguments of the `run` command
+
 ### [1.0.0rc10](https://github.com/ofek/hatch/releases/tag/hatch-v1.0.0rc10) - 2022-01-19 ### {: #hatch-v1.0.0rc10 }
 
 ***Fixed:***